fix(*): make TypeScript realize that each plugin package has a default export
