+++ packages/docusaurus-plugin-client-redirects/src/options.ts
@@ -5,13 +5,34 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import type {
-  PluginOptions,
-  Options,
-  RedirectOption,
-} from '@docusaurus/plugin-client-redirects';
 import type {OptionValidationContext} from '@docusaurus/types';
 import {Joi, PathnameSchema} from '@docusaurus/utils-validation';
+
+export type RedirectOption = {
+  to: string;
+  from: string | string[];
+};
+
+export type PluginOptions = {
+  /** Plugin ID. */
+  id: string;
+  /** The extensions to be removed from the route after redirecting. */
+  fromExtensions: string[];
+  /** The extensions to be appended to the route after redirecting. */
+  toExtensions: string[];
+  /** The list of redirect rules, each one with multiple `from`s → one `to`. */
+  redirects: RedirectOption[];
+  /**
+   * A callback to create a redirect rule.
+   * @returns All the paths from which we should redirect to `path`
+   */
+  createRedirects?: (
+    /** An existing Docusaurus route path */
+    path: string,
+  ) => string[] | string | null | undefined;
+};
+
+export type Options = Partial<PluginOptions>;
 
 export const DEFAULT_OPTIONS: Partial<PluginOptions> = {
   fromExtensions: [],
