+++ packages/docusaurus-plugin-content-blog/src/types.ts
@@ -5,41 +5,9 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import type {BrokenMarkdownLink, ContentPaths, Tag} from '@docusaurus/utils';
-import type {BlogPostMetadata} from '@docusaurus/plugin-content-blog';
-import type {Metadata as BlogPaginatedMetadata} from '@theme/BlogListPage';
+import type {BrokenMarkdownLink, ContentPaths} from '@docusaurus/utils';
 
 export type BlogContentPaths = ContentPaths;
-
-export type BlogContent = {
-  blogSidebarTitle: string;
-  blogPosts: BlogPost[];
-  blogListPaginated: BlogPaginated[];
-  blogTags: BlogTags;
-  blogTagsListPath: string;
-};
-
-export type BlogTags = {
-  [permalink: string]: BlogTag;
-};
-
-export type BlogTag = Tag & {
-  /** Blog post permalinks. */
-  items: string[];
-  pages: BlogPaginated[];
-};
-
-export type BlogPost = {
-  id: string;
-  metadata: BlogPostMetadata;
-  content: string;
-};
-
-export type BlogPaginated = {
-  metadata: BlogPaginatedMetadata;
-  /** Blog post permalinks. */
-  items: string[];
-};
 
 export type BlogBrokenMarkdownLink = BrokenMarkdownLink<BlogContentPaths>;
 export type BlogMarkdownLoaderOptions = {
