+++ packages/docusaurus-plugin-content-docs/src/plugin-content-docs.d.ts
@@ -14,6 +14,7 @@ declare module '@docusaurus/plugin-content-docs' {
     TagModule,
     Tag,
   } from '@docusaurus/utils';
+  import type {Plugin, LoadContext} from '@docusaurus/types';
   import type {Required} from 'utility-types';
 
   export type Assets = {
@@ -504,6 +505,21 @@ declare module '@docusaurus/plugin-content-docs' {
   export type PropTagsListPage = {
     tags: TagsListItem[];
   };
+
+  export type LoadedVersion = VersionMetadata & {
+    docs: DocMetadata[];
+    drafts: DocMetadata[];
+    sidebars: import('./sidebars/types').Sidebars;
+  };
+
+  export type LoadedContent = {
+    loadedVersions: LoadedVersion[];
+  };
+
+  export default function pluginContentDocs(
+    context: LoadContext,
+    options: PluginOptions,
+  ): Promise<Plugin<LoadedContent>>;
 }
 
 declare module '@theme/DocItem' {
