+++ packages/docusaurus-plugin-content-docs/src/types.ts
@@ -7,12 +7,11 @@
 
 /// <reference types="@docusaurus/module-type-aliases" />
 
-import type {Sidebars} from './sidebars/types';
 import type {BrokenMarkdownLink, Tag} from '@docusaurus/utils';
 import type {
   VersionMetadata,
   LastUpdateData,
-  DocMetadata,
+  LoadedVersion,
   CategoryGeneratedIndexMetadata,
 } from '@docusaurus/plugin-content-docs';
 import type {SidebarsUtils} from './sidebars/utils';
@@ -35,16 +34,6 @@ export type VersionTag = Tag & {
 };
 export type VersionTags = {
   [permalink: string]: VersionTag;
-};
-
-export type LoadedVersion = VersionMetadata & {
-  docs: DocMetadata[];
-  drafts: DocMetadata[];
-  sidebars: Sidebars;
-};
-
-export type LoadedContent = {
-  loadedVersions: LoadedVersion[];
 };
 
 export type FullVersion = LoadedVersion & {
