+++ packages/docusaurus-plugin-google-analytics/src/options.ts
