+++ packages/docusaurus-plugin-google-gtag/package.json
@@ -3,7 +3,7 @@
   "version": "2.0.0-beta.18",
   "description": "Global Site Tag (gtag.js) plugin for Docusaurus.",
   "main": "lib/index.js",
-  "types": "src/plugin-google-gtag.d.ts",
+  "types": "lib/index.d.ts",
   "scripts": {
     "build": "tsc",
     "watch": "tsc --watch"
