+++ packages/docusaurus-plugin-google-gtag/src/options.ts
