+++ packages/docusaurus-plugin-sitemap/src/__tests__/options.test.ts
@@ -5,9 +5,8 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import {validateOptions, DEFAULT_OPTIONS} from '../options';
+import {validateOptions, DEFAULT_OPTIONS, type Options} from '../options';
 import {normalizePluginOptions} from '@docusaurus/utils-validation';
-import type {Options} from '@docusaurus/plugin-sitemap';
 
 function testValidate(options: Options) {
   return validateOptions({validate: normalizePluginOptions, options});
