+++ packages/docusaurus-preset-classic/package.json
@@ -3,7 +3,7 @@
   "version": "2.0.0-beta.18",
   "description": "Classic preset for Docusaurus.",
   "main": "lib/index.js",
-  "types": "src/preset-classic.d.ts",
+  "types": "lib/index.d.ts",
   "scripts": {
     "build": "tsc",
     "watch": "tsc --watch"
