+++ packages/docusaurus-preset-classic/src/options.ts
