+++ packages/docusaurus-theme-classic/src/index.ts
@@ -91,10 +91,10 @@ function getInfimaCSSFile(direction: string) {
   }.css`;
 }
 
-export default function docusaurusThemeClassic(
+export default function themeClassic(
   context: LoadContext,
   options: Options,
-): Plugin<void> {
+): Plugin<undefined> {
   const {
     i18n: {currentLocale, localeConfigs},
   } = context;
